+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/items/trinkets/TrapMechanism.java
@@ -37,9 +37,9 @@ public class TrapMechanism extends Trinket {
 	}
 
 	@Override
-	protected int upgradeEnergyCost() {
+	protected long upgradeEnergyCost() {
 		//6 -> 5(11) -> 7(18) -> 8(26)
-		return Math.round(5+1.67f*(int)level());
+		return Math.round(5+1.67f*level());
 	}
 
 	@Override
